+++ packages/cascara/src/structures/AdminStructure/components/Nav/NavLinks.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import pt from 'prop-types';
-import { linkShape } from './Nav';
+import { LINK_SHAPE } from './__propTypes';
 import NavLink from './NavLink';
 
 const propTypes = {
-  links: pt.arrayOf(pt.shape(linkShape)),
+  links: pt.arrayOf(pt.shape(LINK_SHAPE)),
 };
 
 const NavLinks = ({ links }) =>
