+++ packages/cascara/src/structures/AdminStructure/components/Nav/__propTypes.js
@@ -1,0 +1,9 @@
+import pt from 'prop-types';
+
+const LINK_SHAPE = {
+  icon: pt.object,
+  label: pt.string.isRequired,
+  linkComponentProps: pt.object,
+};
+
+export { LINK_SHAPE };
