+++ packages/cascara/src/ui/Table/index.js
@@ -1,1 +1,2 @@
-export { default, propTypes } from './Table';
+export { TABLE_SHAPE as propTypes } from './__propTypes';
+export { default } from './Table';
